refactor: Consolidate profile editing and unified dialog services into `Dialogs/UnifiedProfileDialogService` and remove obsolete dialog, bootloader, and view model factory services.
