+++ src/S7Tools/Services/Core/ApplicationSettingsService.cs
