+++ src/S7Tools/Services/Core/ClipboardService.cs
