+++ src/S7Tools/Services/Core/CommandDispatcher.cs
