+++ src/S7Tools/Services/Core/LocalizationService.cs
