+++ src/S7Tools/Services/Core/PathDiagnosticsService.cs
