+++ src/S7Tools/Services/Core/PathService.cs
