+++ src/S7Tools/Services/Core/ResourceManagerService.cs
