+++ src/S7Tools/Services/Core/ValidationService.cs
