+++ src/S7Tools/Services/Core/WritableOptions.cs
