+++ src/S7Tools/Services/Dialogs/DesignTimeDialogService.cs
