+++ src/S7Tools/Services/Factories/DesignTimeViewModelFactory.cs
