+++ src/S7Tools/Services/Factories/EnhancedViewModelFactory.cs
