+++ src/S7Tools/Services/Factories/TaskLogDataStoreFactory.cs
