+++ src/S7Tools/Services/Interfaces/IProfileDetailsService.cs
