+++ src/S7Tools/Services/Logging/CentralizedTaskLogService.cs
