+++ src/S7Tools/Services/Logging/StructuredLogger.cs
