+++ src/S7Tools/Services/Profiles/MemoryRegionProfileService.cs
