+++ src/S7Tools/Services/Profiles/ProfileDetailsService.cs
