+++ src/S7Tools/Services/Profiles/ProfileValidationResult.cs
