+++ src/S7Tools/Services/SerialPort/SerialPortService.cs
