+++ src/S7Tools/Services/Socat/SocatService.cs
