+++ src/S7Tools/Services/UI/ActivityBarService.cs
