+++ src/S7Tools/Services/UI/ErrorDisplayService.cs
